Reformat the tokenizer into a module.
