+++ src/cli/ast/insert_statement.rs
@@ -1,4 +1,4 @@
-use crate::cli::{ast::interpreter::Interpreter, ast::SqlStatement};
+use crate::cli::{ast::SqlStatement, ast::interpreter::Interpreter};
 
 pub fn build(interpreter: &Interpreter) -> Result<SqlStatement, String> {
     todo!()
