+++ src/cli/mod.rs
@@ -1,18 +1,18 @@
 use std::io;
-mod tokenize;
 mod ast;
 mod table;
+mod tokenizer;
 
 pub fn cli() {
     println!("Welcome to the MollyDB CLI");
     let mut line_count = 1;
-    
+
     loop {
         print!("({:03}) > ", line_count);
         line_count += 1;
-        
+
         io::Write::flush(&mut io::stdout()).unwrap();
-        
+
         let mut input = String::new();
         io::stdin().read_line(&mut input).unwrap();
         let input = input.trim();
@@ -22,9 +22,8 @@ pub fn cli() {
             break;
         }
 
-        let tokens = tokenize::tokenize(input);
+        let tokens = tokenizer::tokenize(input);
         println!("{:?}", tokens);
         ast::generate(tokens);
-
     }
 }
